refactor: make MigrationSnapshotDirector an inner class of SnapshotAfterMigrationTransistionStep
`MigrationSnapshotDirector` is only an utility class used by
`SnapshotAfterMigrationTransistionStep` to ensure a clean context is
used in each transition. In practice, the former contains most of the
logic actually needed for the latter.

Previously the two classes were defined in different files and even in
different packages. This forced various methods to be `public`,
unnecessarily exposing implementation details.

This refactoring moves the utility class inside the main class actually
needing it, ensuring the logic is kept in the same place.

Additionally, this refactoring solves an incorrect use of
`@VisibleForTesting`.
In `SnapshotMigrationTransistionStepTest` there are tests checking if a
snapshot was taken. Previously, this was done by calling the method
`SnapshotAfterMigrationTransitionStep#isSnapshotTaken` which internally
called `MigrationSnapshotDirector#isSnapshotTaken`. Both of these
methods were annotated `@VisibleForTesting` because the `snapshotTaken`
flag is an implementation detail of `MigrationSnapshotDirector`.
This chain of calls is not possible because it requires a
`@VisibleForTesting` method to be called in non-test code.

This refactoring solves this by ensuring that the production code only
accesses those methods as if they were "private".
